# openapi-specification.yml
openapi: 3.0.3
info:
  title: E-Commerce API
  description: 이커머스 플랫폼 API
  version: 1.0.0
  contact:
    name: ehkwon
    email: ehkwon@example.com
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
  - url: https://api.ecommerce.com
    description: 운영 서버
...
paths:
...
  /api/user/{userId}:
    get:
      tags:
        - User
      summary: 유저 조회
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  data:
                    $ref: '#/components/schemas/UserResponse'
                  message:
                    type: string
                    nullable: true
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - User
      summary: 유저 삭제
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 삭제 성공
...
  /api/coupons/user/{userId}:
    get:
      tags:
        - Coupon
      summary: 유저의 쿠폰 목록 조회
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: status
          in: query
          description: 상태 필터 (ACTIVE, USED, EXPIRED)
          schema:
            type: string
            enum: [ACTIVE, USED, EXPIRED]
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserCouponResponse'
  /api/coupons/user/{userId}/{couponId}:
    get:
      tags:
        - Coupon
      summary: 유저의 쿠폰 상세 조회
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: couponId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCouponResponse'
    post:
      tags:
        - Coupon
      summary: 유저에게 일반 쿠폰 발행
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: couponId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          description: 쿠폰 발행 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  data:
                    $ref: '#/components/schemas/UserCouponResponse'
                  message:
                    type: string
                    nullable: true
  /api/coupons/{couponId}/issue:
    patch:
      tags:
        - Coupon
      summary: 선착순 쿠폰 발행
      parameters:
        - name: couponId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  format: int64
      responses:
        '200':
          description: 쿠폰 발행 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    default: true
                  data:
                    $ref: '#/components/schemas/UserCouponResponse'
                  message:
                    type: string
                    nullable: true
        '409':
          description: 쿠폰 소진
...
components:
  schemas:
...
    UserResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        pointBalance:
          type: number
          format: decimal
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UserCouponResponse:
      type: object
      properties:
        userCouponId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        couponId:
          type: integer
          format: int64
        couponName:
          type: string
        discountType:
          type: string
        discountValue:
          type: number
          format: decimal
        status:
          type: string
          enum: [ACTIVE, USED, EXPIRED]
        usedAt:
          type: string
          format: date-time
        orderId:
          type: integer
          format: int64
        issuedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        data:
          type: object
          nullable: true
        message:
          type: string
        errorCode:
          type: string
  responses:
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            data: null
            message: "요청한 리소스를 찾을 수 없습니다"
            errorCode: "RESOURCE_NOT_FOUND"
...